backend fic
I found a problem with the last api. The problem was that the path
arguments was put together with query parameters. I am not an expert of
fast api but in the documentation https://fastapi.tiangolo.com/tutorial/query-params/?h=+quer is
suggested to not combine this two definition

What it's cool to do is to put only path argument inside your api
definiont (ex. @app.get("/mycoolapi/{mycoolpathparameter}"))
and puth the query parameters (the arguments foollowing
? sign) directly in to your api hendler definition.

I also add some dipendencies to the package.json file
I coulden't build the project without
react-redux
styled-components
packages.